Exponentiated Quadratic Covariance

Generate a \(N \times N\) covariance matrix, \(\Sigma\), by passing a distance vector x to the exponentiated quadratic kernel function:

$$\Sigma_{ij} = \alpha^2 \exp \left( \frac{|x_i - x_j|^2}{-2 \ \rho} \right) + \delta_{ij} \ \sigma$$

where \(\alpha^2\) is the amplitude, \(\rho\) is the length_scale, and \(\sigma\) is a small offset along the diagonal to allow for additional variation.

Usage

cov_exp_quad(x, amplitude = 1, length_scale = 1, delta = 1e-09)

Arguments

x

A vector containing positions

amplitude

Numeric, controls the vertical scale of the covariance function

length_scale

Numeric, controls the horizontal scale of the covariance function

delta

A small offset added along the diagonal to ensure that the function returns a positive-semidefinite matrix. Larger values allow for increased variation at individual positions along the vector x.

Value

A \(N \times N\) symmetric, positive-semidefinite covariance matrix

Examples

x <- 1:12
cov_exp_quad(x)
#>               [,1]         [,2]         [,3]         [,4]         [,5]
#>  [1,] 1.000000e+00 6.065307e-01 1.353353e-01 1.110900e-02 3.354626e-04
#>  [2,] 6.065307e-01 1.000000e+00 6.065307e-01 1.353353e-01 1.110900e-02
#>  [3,] 1.353353e-01 6.065307e-01 1.000000e+00 6.065307e-01 1.353353e-01
#>  [4,] 1.110900e-02 1.353353e-01 6.065307e-01 1.000000e+00 6.065307e-01
#>  [5,] 3.354626e-04 1.110900e-02 1.353353e-01 6.065307e-01 1.000000e+00
#>  [6,] 3.726653e-06 3.354626e-04 1.110900e-02 1.353353e-01 6.065307e-01
#>  [7,] 1.522998e-08 3.726653e-06 3.354626e-04 1.110900e-02 1.353353e-01
#>  [8,] 2.289735e-11 1.522998e-08 3.726653e-06 3.354626e-04 1.110900e-02
#>  [9,] 1.266417e-14 2.289735e-11 1.522998e-08 3.726653e-06 3.354626e-04
#> [10,] 2.576757e-18 1.266417e-14 2.289735e-11 1.522998e-08 3.726653e-06
#> [11,] 1.928750e-22 2.576757e-18 1.266417e-14 2.289735e-11 1.522998e-08
#> [12,] 5.311092e-27 1.928750e-22 2.576757e-18 1.266417e-14 2.289735e-11
#>               [,6]         [,7]         [,8]         [,9]        [,10]
#>  [1,] 3.726653e-06 1.522998e-08 2.289735e-11 1.266417e-14 2.576757e-18
#>  [2,] 3.354626e-04 3.726653e-06 1.522998e-08 2.289735e-11 1.266417e-14
#>  [3,] 1.110900e-02 3.354626e-04 3.726653e-06 1.522998e-08 2.289735e-11
#>  [4,] 1.353353e-01 1.110900e-02 3.354626e-04 3.726653e-06 1.522998e-08
#>  [5,] 6.065307e-01 1.353353e-01 1.110900e-02 3.354626e-04 3.726653e-06
#>  [6,] 1.000000e+00 6.065307e-01 1.353353e-01 1.110900e-02 3.354626e-04
#>  [7,] 6.065307e-01 1.000000e+00 6.065307e-01 1.353353e-01 1.110900e-02
#>  [8,] 1.353353e-01 6.065307e-01 1.000000e+00 6.065307e-01 1.353353e-01
#>  [9,] 1.110900e-02 1.353353e-01 6.065307e-01 1.000000e+00 6.065307e-01
#> [10,] 3.354626e-04 1.110900e-02 1.353353e-01 6.065307e-01 1.000000e+00
#> [11,] 3.726653e-06 3.354626e-04 1.110900e-02 1.353353e-01 6.065307e-01
#> [12,] 1.522998e-08 3.726653e-06 3.354626e-04 1.110900e-02 1.353353e-01
#>              [,11]        [,12]
#>  [1,] 1.928750e-22 5.311092e-27
#>  [2,] 2.576757e-18 1.928750e-22
#>  [3,] 1.266417e-14 2.576757e-18
#>  [4,] 2.289735e-11 1.266417e-14
#>  [5,] 1.522998e-08 2.289735e-11
#>  [6,] 3.726653e-06 1.522998e-08
#>  [7,] 3.354626e-04 3.726653e-06
#>  [8,] 1.110900e-02 3.354626e-04
#>  [9,] 1.353353e-01 1.110900e-02
#> [10,] 6.065307e-01 1.353353e-01
#> [11,] 1.000000e+00 6.065307e-01
#> [12,] 6.065307e-01 1.000000e+00