Reformat Memorial Hermann Survey Names

Convenience function for reformatting MH Survey names as they appear in the Qualtrics UI to something more legible.

Usage

fix_survey_names(.data, survey_names)

Arguments

.data

A tibble or dataframe containing survey names.

survey_names

The name of the column in .data containing survey names.

Examples

# list of survey names as they appear in the qualtrics ui
surveys <-
  tibble::tibble(
    names = c("Live - Adult Day Surgery",
              "Live - Adult Emergency Department",
              "Live - Adult Inpatient Rehab",
              "Live - Adult Medical Practice",
              "Live - Adult Outpatient",
              "Live - Adult Outpatient - Oncology",
              "Live - Adult Outpatient - Vaccine",
              "Live - Adult Outpatient Rehab",
              "Live - Adult Telemedicine",
              "Live - Adult Urgent Care",
              "Live - HCAHPS - Paper",
              "Live - Home Medical Equipment",
              "Live - Infusion Pharmacy",
              "Live - Inpatient",
              "Live - Pediatric - Emergency Department",
              "Live - Pediatric Inpatient")
  )

# bad names!
surveys
#> # A tibble: 16 × 1
#>    names                                  
#>    <chr>                                  
#>  1 Live - Adult Day Surgery               
#>  2 Live - Adult Emergency Department      
#>  3 Live - Adult Inpatient Rehab           
#>  4 Live - Adult Medical Practice          
#>  5 Live - Adult Outpatient                
#>  6 Live - Adult Outpatient - Oncology     
#>  7 Live - Adult Outpatient - Vaccine      
#>  8 Live - Adult Outpatient Rehab          
#>  9 Live - Adult Telemedicine              
#> 10 Live - Adult Urgent Care               
#> 11 Live - HCAHPS - Paper                  
#> 12 Live - Home Medical Equipment          
#> 13 Live - Infusion Pharmacy               
#> 14 Live - Inpatient                       
#> 15 Live - Pediatric - Emergency Department
#> 16 Live - Pediatric Inpatient             

# better names!
fix_survey_names(surveys, names)
#> # A tibble: 16 × 1
#>    names                 
#>    <chr>                 
#>  1 Day Surgery           
#>  2 Emergency             
#>  3 Inpatient Rehab       
#>  4 Medical Practice      
#>  5 Outpatient            
#>  6 Outpatient Oncology   
#>  7 Outpatient Vaccine    
#>  8 Outpatient Rehab      
#>  9 Telemedicine          
#> 10 Urgent Care           
#> 11 HCAHPS                
#> 12 Home Medical Equipment
#> 13 Infusion Pharmacy     
#> 14 Inpatient             
#> 15 Pedi Emergency        
#> 16 Pedi Inpatient