Transform a between bound and unbounded scales.

Transform a between bound and unbounded scales.

Usage

logit(x, lower_bound = 0, upper_bound = 1, base = exp(1))

expit(x, lower_bound = 0, upper_bound = 1, base = exp(1))

Arguments

x

Value to be rescaled.

lower_bound

Lower bound of x; defaults to 0.

upper_bound

Upper bound of x; defaults to 1.

base

Base used for log-transform, defaults to exp(1) = e

Details

logit() will transform a value on a bound scale to an unbound scale, while expit() performs the reverse transformation. Values at the lower/upper bounds correspond to -Inf/Inf on the unbounded scale. Uses the natural base, e, by default, but other bases can be passed to the parameter base.

Examples

# unbound vector of values bound between [0, 1]
myvec <- seq(0, 10, 1)/10
logit(myvec)
#>  [1]       -Inf -2.1972246 -1.3862944 -0.8472979 -0.4054651  0.0000000
#>  [7]  0.4054651  0.8472979  1.3862944  2.1972246        Inf

# unbound vector of values between [100, 200]
myvec <- seq(100, 200, 10)
logit(myvec, lower_bound = 100, upper_bound = 200)
#>  [1]       -Inf -2.1972246 -1.3862944 -0.8472979 -0.4054651  0.0000000
#>  [7]  0.4054651  0.8472979  1.3862944  2.1972246        Inf

# bound vector of unbounded values to [0, 1]
myvec <- rnorm(10)
expit(myvec)
#>  [1] 0.4678692 0.4667147 0.1114772 0.2557297 0.5845928 0.7040486 0.4832242
#>  [8] 0.7389116 0.2238045 0.2761972

# bound vector of unbounded values to [100, 200]
expit(myvec, lower_bound = 100, upper_bound = 200)
#>  [1] 146.7869 146.6715 111.1477 125.5730 158.4593 170.4049 148.3224 173.8912
#>  [9] 122.3805 127.6197