Apply a confidence interval to values in a dataframe based on the normal distribution.

Apply a confidence interval to values in a dataframe based on the normal distribution.

Usage

normal_interval(.data, mean, std_dev, conf = 0.95)

Arguments

.data

Data frame; a confidence interval will be applied to each row in the data frame.

mean

Mean of the distribution

std_dev

Standard deviation of the distribution

conf

Confidence interval, must be between [0, 1]. Defaults to 0.95.

See also

Examples

# create example df
mean <- rnorm(10, 100, 20)
std_dev <- rnorm(10, 80, 10)
example <- dplyr::bind_cols(mean, std_dev)
#> New names:
#> • `` -> `...1`
#> • `` -> `...2`
colnames(example) <- c("mean", "std_dev")

# apply the default confidence interval of 0.95
normal_interval(example, mean, std_dev)
#> # A tibble: 10 × 4
#>     mean std_dev ci_lower ci_upper
#>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 108.     73.6    -36.3     252.
#>  2 134.     79.8    -22.6     290.
#>  3 108.     78.6    -46.1     262.
#>  4  92.1    74.5    -53.8     238.
#>  5 103.     83.0    -59.3     266.
#>  6 116.     89.6    -59.2     292.
#>  7  76.7    71.6    -63.7     217.
#>  8  98.7    68.4    -35.4     233.
#>  9  97.4    96.4    -91.6     286.
#> 10 135.     78.1    -17.9     288.

# apply a custom confidence interval
normal_interval(example, mean, std_dev, conf = 0.99)
#> # A tibble: 10 × 4
#>     mean std_dev ci_lower ci_upper
#>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1 108.     73.6    -81.6     298.
#>  2 134.     79.8    -71.8     339.
#>  3 108.     78.6    -94.6     311.
#>  4  92.1    74.5    -99.7     284.
#>  5 103.     83.0   -110.      317.
#>  6 116.     89.6   -114.      347.
#>  7  76.7    71.6   -108.      261.
#>  8  98.7    68.4    -77.5     275.
#>  9  97.4    96.4   -151.      346.
#> 10 135.     78.1    -66.0     336.